+++ Content/Items/Reds/ResonanceDisc.cs
@@ -45,6 +45,33 @@
             var projectileImpactExplosion = bomb.GetComponent<ProjectileImpactExplosion>();
             projectileImpactExplosion.falloffModel = BlastAttack.FalloffModel.None;
             projectileImpactExplosion.blastRadius = explosionRadius;
+            projectileImpactExplosion.blastProcCoefficient = explosionProcCoefficient * globalProc;
+
+            var explosion = PrefabAPI.InstantiateClone(Utils.Paths.GameObject.LaserTurbineBombExplosion.Load<GameObject>(), "Resonance Disc Explosion", false);
+            var transform = explosion.transform;
+            for (int i = 0; i < transform.childCount; i++)
+            {
+                var index = transform.GetChild(i);
+                index.transform.localScale *= 2f;
+            }
+
+            var sparklers = transform.GetChild(3);
+            for (int j = 0; j < sparklers.transform.childCount; j++)
+            {
+                var index = sparklers.transform.GetChild(j);
+                index.transform.localScale *= 2f;
+            }
+
+            var evis = transform.GetChild(5);
+            for (int k = 0; k < evis.transform.childCount; k++)
+            {
+                var index = evis.transform.GetChild(k);
+                index.transform.localScale *= 2f;
+            }
+
+            ContentAddition.AddEffect(explosion);
+
+            projectileImpactExplosion.impactEffect = explosion;
         }
     }
 }